+++ src/Console/LighthouseGeneratorCommand.php
@@ -3,6 +3,7 @@
 namespace Nuwave\Lighthouse\Console;
 
 use Illuminate\Console\GeneratorCommand;
+use InvalidArgumentException;
 
 abstract class LighthouseGeneratorCommand extends GeneratorCommand
 {
@@ -51,34 +52,45 @@ abstract class LighthouseGeneratorCommand extends GeneratorCommand
     public static function commonNamespace(array $namespaces): string
     {
         if ($namespaces === []) {
-            return '';
+            throw new InvalidArgumentException(
+                'A default namespace is required for code generation.'
+            );
         }
 
         if (count($namespaces) === 1) {
             return reset($namespaces);
         }
 
+        // Save the first namespac
+        $preferredNamespaceFallback = reset($namespaces);
+
         // If the strings are sorted, any prefix common to all strings
         // will be common to the sorted first and last strings.
         // All the strings in the middle can be ignored.
         sort($namespaces);
 
-        $first = explode('\\', reset($namespaces));
-        $last = explode('\\', end($namespaces));
+        $firstParts = explode('\\', reset($namespaces));
+        $lastParts = explode('\\', end($namespaces));
 
         $matching = [];
-        foreach ($first as $i => $part) {
+        foreach ($firstParts as $i => $part) {
             // We ran out of elements to compare, so we reached the maximum common length
-            if (! isset($last[$i])) {
+            if (! isset($lastParts[$i])) {
                 break;
             }
 
             // We found an element that differs
-            if ($last[$i] !== $part) {
+            if ($lastParts[$i] !== $part) {
                 break;
             }
 
             $matching [] = $part;
+        }
+
+        // We could not determine a common part of the configured namespaces,
+        // so we just assume the user will prefer the first one in the list.
+        if ($matching === []) {
+            return $preferredNamespaceFallback;
         }
 
         return implode('\\', $matching);