+++ src/Execution/Arguments/ArgumentSet.php
@@ -6,7 +6,6 @@ use Closure;
 use Nuwave\Lighthouse\Schema\Directives\RenameDirective;
 use Nuwave\Lighthouse\Schema\Directives\SpreadDirective;
 use Nuwave\Lighthouse\Support\Contracts\ArgBuilderDirective;
-use Nuwave\Lighthouse\Support\Contracts\Directive;
 use Nuwave\Lighthouse\Support\Utils;
 
 class ArgumentSet
@@ -82,9 +81,7 @@ class ArgumentSet
                 $value = $value->spread();
 
                 if ($argument->directives->contains(
-                    function (Directive $directive): bool {
-                        return $directive instanceof SpreadDirective;
-                    }
+                    Utils::instanceofMatcher(SpreadDirective::class)
                 )) {
                     $argumentSet->arguments += $value->arguments;
                     continue;
@@ -180,9 +177,7 @@ class ArgumentSet
 
             $filteredDirectives = $argument
                 ->directives
-                ->filter(function (Directive $directive): bool {
-                    return $directive instanceof ArgBuilderDirective;
-                });
+                ->filter(Utils::instanceofMatcher(ArgBuilderDirective::class));
 
             if (! empty($directiveFilter)) {
                 $filteredDirectives = $filteredDirectives->filter($directiveFilter);