+++ src/LighthouseServiceProvider.php
@@ -15,6 +15,7 @@ use Illuminate\Validation\Factory as ValidationFactory;
 use Illuminate\Validation\Validator;
 use Laravel\Lumen\Application as LumenApplication;
 use Nuwave\Lighthouse\Console\ClearCacheCommand;
+use Nuwave\Lighthouse\Console\DirectiveCommand;
 use Nuwave\Lighthouse\Console\IdeHelperCommand;
 use Nuwave\Lighthouse\Console\InterfaceCommand;
 use Nuwave\Lighthouse\Console\MutationCommand;
@@ -39,6 +40,7 @@ use Nuwave\Lighthouse\Schema\Source\SchemaSourceProvider;
 use Nuwave\Lighthouse\Schema\Source\SchemaStitcher;
 use Nuwave\Lighthouse\Schema\TypeRegistry;
 use Nuwave\Lighthouse\Schema\Values\FieldValue;
+use Nuwave\Lighthouse\Support\AppVersion;
 use Nuwave\Lighthouse\Support\Compatibility\LaravelMiddlewareAdapter;
 use Nuwave\Lighthouse\Support\Compatibility\LumenMiddlewareAdapter;
 use Nuwave\Lighthouse\Support\Compatibility\MiddlewareAdapter;
@@ -49,6 +51,7 @@ use Nuwave\Lighthouse\Support\Contracts\GlobalId as GlobalIdContract;
 use Nuwave\Lighthouse\Support\Contracts\ProvidesResolver;
 use Nuwave\Lighthouse\Support\Contracts\ProvidesSubscriptionResolver;
 use Nuwave\Lighthouse\Support\Http\Responses\ResponseStream;
+use Nuwave\Lighthouse\Testing\TestingServiceProvider;
 
 class LighthouseServiceProvider extends ServiceProvider
 {
@@ -89,7 +92,7 @@ class LighthouseServiceProvider extends ServiceProvider
      */
     protected function loadRoutesFrom($path): void
     {
-        if (Str::contains($this->app->version(), 'Lumen')) {
+        if (AppVersion::isLumen()) {
             require realpath($path);
 
             return;
@@ -169,6 +172,7 @@ class LighthouseServiceProvider extends ServiceProvider
         if ($this->app->runningInConsole()) {
             $this->commands([
                 ClearCacheCommand::class,
+                DirectiveCommand::class,
                 IdeHelperCommand::class,
                 InterfaceCommand::class,
                 MutationCommand::class,
@@ -179,6 +183,10 @@ class LighthouseServiceProvider extends ServiceProvider
                 UnionCommand::class,
                 ValidateSchemaCommand::class,
             ]);
+        }
+
+        if ($this->app->runningUnitTests()) {
+            $this->app->register(TestingServiceProvider::class);
         }
     }
 }