+++ src/Support/Utils.php
@@ -95,4 +95,32 @@ class Utils
 
         return array_map($callback, $valueOrValues);
     }
+
+    /**
+     * Determine if a class uses a trait.
+     *
+     * @param  object|string  $class
+     * @param  string  $trait
+     * @return bool
+     */
+    public static function classUsesTrait($class, string $trait): bool
+    {
+        return in_array(
+            $trait,
+            class_uses_recursive($class)
+        );
+    }
+
+    /**
+     * Construct a callback that checks if its input is a given class.
+     *
+     * @param  string  $classLike
+     * @return \Closure
+     */
+    public static function instanceofMatcher(string $classLike): \Closure
+    {
+        return function ($object) use ($classLike): bool {
+            return $object instanceof $classLike;
+        };
+    }
 }