+++ src/WhereConditions/WhereConditionsServiceProvider.php
@@ -8,7 +8,6 @@ use Illuminate\Support\ServiceProvider;
 use Nuwave\Lighthouse\Events\ManipulateAST;
 use Nuwave\Lighthouse\Events\RegisterDirectiveNamespaces;
 use Nuwave\Lighthouse\Schema\AST\PartialParser;
-use Nuwave\Lighthouse\Schema\Factories\DirectiveFactory;
 
 class WhereConditionsServiceProvider extends ServiceProvider
 {
@@ -27,11 +26,10 @@ class WhereConditionsServiceProvider extends ServiceProvider
     /**
      * Bootstrap any application services.
      *
-     * @param  \Nuwave\Lighthouse\Schema\Factories\DirectiveFactory  $directiveFactory
      * @param  \Illuminate\Contracts\Events\Dispatcher  $dispatcher
      * @return void
      */
-    public function boot(DirectiveFactory $directiveFactory, Dispatcher $dispatcher): void
+    public function boot(Dispatcher $dispatcher): void
     {
         $dispatcher->listen(
             RegisterDirectiveNamespaces::class,