+++ tests/Integration/GraphQLTest.php
@@ -153,7 +153,8 @@ class GraphQLTest extends TestCase
                         'message' => $message,
                     ],
                 ],
-            ]);
+            ])
+            ->assertStatus(200);
     }
 
     public function testIgnoresInvalidJSONVariables(): void