+++ tests/Integration/ValidationTest.php
@@ -2,7 +2,6 @@
 
 namespace Tests\Integration;
 
-use Illuminate\Foundation\Testing\TestResponse;
 use Illuminate\Support\Arr;
 use Illuminate\Support\Str;
 use Tests\DBTestCase;
@@ -29,7 +28,7 @@ class ValidationTest extends DBTestCase
             password: String
                 @trim
                 @rules(apply: ["min:6", "max:20", "required_with:id"])
-                @bcrypt
+                @hash
             bar: Bar
                 @rules(apply: ["required_if:id,bar"])
         ): String @field(resolver: "Tests\\\\Integration\\\\ValidationTest@resolvePassword")
@@ -530,10 +529,10 @@ class ValidationTest extends DBTestCase
      * Assert that the returned result contains an exactly defined array of validation keys.
      *
      * @param  array  $keys
-     * @param  \Illuminate\Foundation\Testing\TestResponse  $result
+     * @param  \Illuminate\Foundation\Testing\TestResponse|\Illuminate\Testing\TestResponse  $result
      * @return void
      */
-    protected function assertValidationKeysSame(array $keys, TestResponse $result): void
+    protected function assertValidationKeysSame(array $keys, $result): void
     {
         $validation = $result->jsonGet('errors.0.extensions.validation');